Add files via upload
Story and workflow for the Treasure Hunt:

The player is a pirate and is looking for the treasure as our team name is Pirates it adds up. The pirate ship will travel from one port to another using the clue in his current state. This is a deterministic finite automata as there is only one correct path.

I am at port of Baltimore and  the next state is port of Miami -dade. 
The Screen will show a hint “Cruise Capital of the World”.
We will also be show pictures of 4 ports. When you hover over the pictures you can read the ports name. If you drag the hint towards the correct picture your state changes and you go to the next state i.e port of Miami. If you drag towards wrong option your ship gets hit by a canon and you will be promoted another hint for the same next state. If you get hit for 5 times in any state you lose. If you manage to complete the hunt and reach the treasure in the least possible time you win. If 5 people are playing simultaneously the one who reaches 1st wins.