Add files via upload
Simple Quiz.
At state(n)
Question is displayed on screen with options. 
If the answer is correct a bonus question appears with option. If answer is correct you advance to state (n+2). If you get the bonus wrong then you advance to state(n+1).
If you answer the first question wrong you will be given another question but in the same state. You can give a total of 5 wrong answers in any state.
The time taken to complete the quiz will be noted. Person with the least time will win.